+++ src/app/features/movie-details/movie-details.spec.ts
@@ -3,28 +3,12 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MovieDetails } from './movie-details';
 import { provideHttpClient } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
-import { of } from 'rxjs';
+import { Router } from '@angular/router';
 import { VotePercentPipe } from '../../shared/pipes/vote-percent-pipe';
 
 describe('MovieDetails', () => {
   let component: MovieDetails;
   let fixture: ComponentFixture<MovieDetails>;
-
-  const mockParameterMap: ParamMap = {
-    get: (key: string) => {
-      if (key === 'id') return '1';
-      if (key === 'type') return 'tv';
-      return null;
-    },
-    getAll: (key: string) => {
-      void key;
-      return [];
-    },
-
-    has: () => false,
-    keys: [],
-  };
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -32,18 +16,6 @@ describe('MovieDetails', () => {
       providers: [
         provideHttpClient(),
         provideHttpClientTesting(),
-        {
-          provide: ActivatedRoute,
-          useValue: {
-            paramMap: of(mockParameterMap),
-            queryParamMap: of({ get: () => null }),
-            snapshot: {
-              paramMap: mockParameterMap,
-              queryParamMap: { get: () => null },
-            },
-          },
-        },
-
         {
           provide: Router,
           useValue: {
@@ -55,6 +27,8 @@ describe('MovieDetails', () => {
 
     fixture = TestBed.createComponent(MovieDetails);
     component = fixture.componentInstance;
+    fixture.componentRef.setInput('id', 1);
+    fixture.componentRef.setInput('type', 'movie');
 
     fixture.detectChanges();
   });