+++ src/app/features/movie-details/movie-details.ts
@@ -1,41 +1,43 @@
-import { ChangeDetectionStrategy, Component, inject, signal } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ChangeDetectionStrategy, Component, inject, signal, input, computed } from '@angular/core';
 import { MediaType } from './types/media-type';
-import { toSignal } from '@angular/core/rxjs-interop';
+import { toObservable, toSignal } from '@angular/core/rxjs-interop';
 import { DetailsCardService } from './services/details-card-service';
 import { MovieDetailsCard } from './movie-details-card/movie-details-card';
-import { SPINNER_PATH } from '../../shared/constants/constants';
 import { HorizontalSlider } from '../../shared/ui/horizontal-slider/horizontal-slider';
 import { TopBilledCastService } from './services/top-billed-cast-service';
 import { PersonCard } from '../../shared/ui/person-card/person-card';
 import { RecommendationsService } from './services/recommendations-service';
 import { RecommendationCard } from './recommendation-card/recommendation-card';
-import { BehaviorSubject, catchError, map, of, switchMap, tap } from 'rxjs';
+import { catchError, of, switchMap, tap } from 'rxjs';
+import { Spinner } from '../../shared/ui/spinner/spinner';
 
 @Component({
   selector: 'app-movie-details',
-  imports: [MovieDetailsCard, HorizontalSlider, PersonCard, RecommendationCard],
+  imports: [MovieDetailsCard, HorizontalSlider, PersonCard, RecommendationCard, Spinner],
   templateUrl: './movie-details.html',
   styleUrl: './movie-details.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class MovieDetails {
-  readonly route = inject(ActivatedRoute);
+  readonly id = input.required<number>();
+  readonly type = input.required<MediaType>();
+
   readonly detailsCardsService = inject(DetailsCardService);
   readonly topBilledCastService = inject(TopBilledCastService);
   readonly recommendationsService = inject(RecommendationsService);
-  readonly spinnerPath = SPINNER_PATH;
 
-  private readonly routeParams$ = new BehaviorSubject<{ id: number; type: MediaType }>({
-    id: Number(this.route.snapshot.paramMap.get('id')),
-    type: (this.route.snapshot.paramMap.get('type') ?? 'movie') as MediaType,
-  });
+  private readonly routeParams = computed(() => ({
+    id: this.id(),
+    type: this.type(),
+  }));
+
+  private readonly parameters$ = toObservable(this.routeParams);
 
   readonly cardDetailsError = this.detailsCardsService.errorSignal;
   readonly cardDetailsLoading = signal(true);
 
   readonly cardDetails = toSignal(
-    this.routeParams$.pipe(
+    this.parameters$.pipe(
       tap(() => {
         this.cardDetailsLoading.set(true);
       }),
@@ -52,13 +54,13 @@ export class MovieDetails {
   readonly castLoading = signal(true);
 
   readonly cast = toSignal(
-    this.routeParams$.pipe(
+    this.parameters$.pipe(
       tap(() => {
         this.castLoading.set(true);
         this.castError.set(null);
       }),
-      switchMap(({ id }) =>
-        this.topBilledCastService.getCast(id).pipe(
+      switchMap(({ id, type }) =>
+        this.topBilledCastService.getCast(id, type).pipe(
           tap({
             next: () => this.castLoading.set(false),
             error: () => this.castLoading.set(false),
@@ -74,12 +76,12 @@ export class MovieDetails {
   readonly recommendationsLoading = signal(true);
 
   readonly recommendations = toSignal(
-    this.routeParams$.pipe(
+    this.parameters$.pipe(
       tap(() => {
         this.recommendationsLoading.set(true);
       }),
-      switchMap(({ id }) =>
-        this.recommendationsService.getRecommendations(id).pipe(
+      switchMap(({ id, type }) =>
+        this.recommendationsService.getRecommendations(id, type).pipe(
           tap({
             next: () => this.recommendationsLoading.set(false),
             error: () => this.recommendationsLoading.set(false),
@@ -90,15 +92,4 @@ export class MovieDetails {
     ),
     { initialValue: [] },
   );
-
-  constructor() {
-    this.route.paramMap
-      .pipe(
-        map((parameters) => ({
-          id: Number(parameters.get('id')),
-          type: (parameters.get('type') ?? 'movie') as MediaType,
-        })),
-      )
-      .subscribe(this.routeParams$);
-  }
 }