+++ src/app/features/person-details/casted-in-card.interface.ts
@@ -3,4 +3,5 @@ export interface CastedInCard {
   title: string;
   name?: string;
   poster_path: string;
+  media_type: 'movie' | 'tv';
 }