+++ src/app/features/person-details/person-details.spec.ts
@@ -1,28 +1,13 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PersonDetails } from './person-details';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
+import { Router } from '@angular/router';
 import { provideHttpClient } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
-import { of } from 'rxjs';
 
 describe('PersonDetails', () => {
   let component: PersonDetails;
   let fixture: ComponentFixture<PersonDetails>;
-
-  const mockParameterMap: ParamMap = {
-    get: (key: string) => {
-      if (key === 'id') return '1';
-      return null;
-    },
-    getAll: (key: string) => {
-      void key;
-      return [];
-    },
-
-    has: () => false,
-    keys: [],
-  };
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -30,18 +15,6 @@ describe('PersonDetails', () => {
       providers: [
         provideHttpClient(),
         provideHttpClientTesting(),
-        {
-          provide: ActivatedRoute,
-          useValue: {
-            paramMap: of(mockParameterMap),
-            queryParamMap: of({ get: () => null }),
-            snapshot: {
-              paramMap: mockParameterMap,
-              queryParamMap: { get: () => null },
-            },
-          },
-        },
-
         {
           provide: Router,
           useValue: {
@@ -53,6 +26,8 @@ describe('PersonDetails', () => {
 
     fixture = TestBed.createComponent(PersonDetails);
     component = fixture.componentInstance;
+    fixture.componentRef.setInput('id', 1);
+
     fixture.detectChanges();
   });