+++ src/app/features/person-details/person-details.ts
@@ -4,30 +4,31 @@ import {
   computed,
   effect,
   inject,
+  input,
   signal,
 } from '@angular/core';
 import { PersonDetailsService } from './services/person-details-service';
-import { ActivatedRoute, RouterLink } from '@angular/router';
-import { cardTrailerURL, FALLBACK_ACTOR, SPINNER_PATH } from '../../shared/constants/constants';
-import { BehaviorSubject } from 'rxjs';
+import { RouterLink } from '@angular/router';
+import { cardTrailerURL, FALLBACK_ACTOR } from '../../shared/constants/constants';
 import { PersonDetailsItem } from '../../pages/models/people/person-details.interface';
 
 import { CastedInService } from './services/casted-in-service';
 import { HorizontalSlider } from '../../shared/ui/horizontal-slider/horizontal-slider';
 import { CastedInCard } from './casted-in-card.interface';
+import { Spinner } from '../../shared/ui/spinner/spinner';
 
 @Component({
   selector: 'app-person-details',
-  imports: [HorizontalSlider, RouterLink],
+  imports: [HorizontalSlider, RouterLink, Spinner],
   templateUrl: './person-details.html',
   styleUrl: './person-details.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PersonDetails {
+  readonly id = input.required<number>();
   readonly personDetailsService = inject(PersonDetailsService);
   readonly castedInService = inject(CastedInService);
-  readonly route = inject(ActivatedRoute);
-  readonly spinnerPath = SPINNER_PATH;
+
   readonly detailsError = signal<string | null>(null);
   readonly castedInError = signal<string | null>(null);
   readonly personDetails = signal<PersonDetailsItem | null>(null);
@@ -37,10 +38,6 @@ export class PersonDetails {
   readonly cardImg = cardTrailerURL;
 
   protected imageUrl = cardTrailerURL;
-
-  private readonly routeParams$ = new BehaviorSubject<{ id: number }>({
-    id: Number(this.route.snapshot.paramMap.get('id')),
-  });
 
   protected uiState = computed(() => {
     if (this.loadingDetails()) return 'loading';
@@ -58,21 +55,13 @@ export class PersonDetails {
   protected isBioExpanded = false;
 
   constructor() {
-    this.route.paramMap.subscribe((parameters) => {
-      const rawId = parameters.get('id');
-      const id = rawId ? Number(rawId) : 0;
-      console.log('Route changed, new id:', id);
-      this.routeParams$.next({ id });
-    });
-
     effect(() => {
-      const { id } = this.routeParams$.value;
-      if (!id) return;
+      if (!this.id()) return;
 
       this.loadingDetails.set(true);
       this.detailsError.set(null);
 
-      this.personDetailsService.getPersonDetails(id).subscribe({
+      const subscription = this.personDetailsService.getPersonDetails(this.id()).subscribe({
         next: (details) => {
           this.personDetails.set(details);
           this.loadingDetails.set(false);
@@ -82,16 +71,16 @@ export class PersonDetails {
           this.loadingDetails.set(false);
         },
       });
+      return () => subscription.unsubscribe();
     });
 
     effect(() => {
-      const { id } = this.routeParams$.value;
-      if (!id) return;
+      if (!this.id()) return;
 
       this.loadingCastedIn.set(true);
       this.castedInError.set(null);
 
-      this.castedInService.getCastedIn(id).subscribe({
+      const subscription = this.castedInService.getCastedIn(this.id()).subscribe({
         next: (items) => {
           this.castedIn.set(items);
           this.loadingCastedIn.set(false);
@@ -101,6 +90,7 @@ export class PersonDetails {
           this.loadingCastedIn.set(false);
         },
       });
+      return () => subscription.unsubscribe();
     });
   }