+++ src/app/pages/home/slider-card/slider-card.spec.ts
@@ -26,6 +26,7 @@ describe('SliderCard', () => {
       release_date: '2025-09-09',
       vote_average: 8.5,
       vote_count: 100,
+      media_type: 'movie',
     } satisfies MovieCard);
     fixture.detectChanges();
   });