+++ src/app/shared/constants/constants.ts
@@ -10,7 +10,6 @@ export const posterURL = 'https://image.tmdb.org/t/p/w1280';
 export const youtubeWatchUrl = 'https://www.youtube.com/watch?v=';
 export const cardTrailerURL = 'https://image.tmdb.org/t/p/w500';
 export const actorPhotoURL = 'https://image.tmdb.org/t/p/w300';
-export const SPINNER_PATH = 'spinner.png';
 export const FALLBACK_TITLE = 'Untitled';
 export const FALLBACK_DATE = 'No date';
 export const FALLBACK_POSTER = 'placeholder-movie.png';