+++ src/app/shared/header/menu/services/menu-service.ts
@@ -7,20 +7,22 @@ import { NO_INDEX } from '../../../constants/constants';
 })
 export class MenuService {
   items: MenuItem[] = [];
-  showVerticalMenu = false;
+  showVerticalMenu = signal(false);
   readonly hoveredIndex = signal<number>(NO_INDEX);
   readonly hoveredSubmenu = signal<MenuItem[] | []>([]);
   readonly mouseInItem = signal(false);
   readonly mouseInPopup = signal(false);
   readonly clickedIndex = signal<number>(NO_INDEX);
   menuIsHovered = signal(false);
+  burgerClass = signal<'open' | 'close' | ''>('');
 
   toggleMenu(): void {
-    if (!this.showVerticalMenu) {
+    if (!this.showVerticalMenu()) {
       this.clickedIndex.set(NO_INDEX);
       this.hoveredSubmenu.set([]);
     }
-    this.showVerticalMenu = !this.showVerticalMenu;
+    this.showVerticalMenu.update((value) => !value);
+    this.burgerClass.set(this.showVerticalMenu() ? 'open' : 'close');
   }
 
   resetSubmenu(): void {
@@ -28,6 +30,7 @@ export class MenuService {
     this.hoveredSubmenu.set([]);
     this.mouseInItem.set(false);
     this.mouseInPopup.set(false);
-    this.showVerticalMenu = false;
+    this.showVerticalMenu.set(false);
+    this.burgerClass.update((value) => (value === 'open' ? 'close' : ''));
   }
 }